+++ qt4-gui/src/views/userview.cpp
@@ -118,7 +118,7 @@ void UserView::updateRootIndex()
 void UserView::configUpdated()
 {
   // Set column widths
-  for (unsigned short i = 0; i < Config::ContactList::instance()->columnCount(); i++)
+  for (int i = 0; i < Config::ContactList::instance()->columnCount(); i++)
     setColumnWidth(i, Config::ContactList::instance()->columnWidth(i));
 
   setVerticalScrollBarPolicy(Config::ContactList::instance()->allowScrollBar() ?
